Should work!
The lab is finsihed to the completion that the prof wants but I'd like
to figure out a way to have it so that if after moving to the end of the
screen the text dissapears instead of the text only dissapearing on
reset. Not sure how I'd do that.